Add new decorator for projects
