+++ asreview/webapp/api/projects.py
@@ -58,10 +58,8 @@ from asreview.models.query import list_query_strategies
 from asreview.project import ASReviewProject
 from asreview.project import ProjectNotFoundError
 from asreview.project import get_project_path
-from asreview.project import get_projects
 from asreview.project import is_v0_project
 from asreview.project import open_state
-from asreview.project import project_from_id
 from asreview.search import SearchError
 from asreview.search import fuzzy_find
 from asreview.settings import ASReviewSettings
@@ -75,7 +73,7 @@ from asreview.utils import asreview_path
 from asreview.utils import list_reader_names
 from asreview.webapp import DB
 from flask_login import login_required
-from asreview.webapp.authentication.login_required import project_authorization
+from asreview.webapp.authentication.decorators import project_authorization, current_user_projects
 from asreview.webapp.authentication.models import Project
 from asreview.webapp.io import read_data
 
@@ -114,24 +112,16 @@ def error_500(e):
 # routes
 @bp.route("/projects", methods=["GET"])
 @login_required
-def api_get_projects():  # noqa: F401
+@current_user_projects
+def api_get_projects(projects):  # noqa: F401
     """"""
     project_info = []
 
-    if current_app.config.get("LOGIN_DISABLED", False):
-        projects = get_projects(None)
-        owner_ids = [None for p in projects]
-    else:
-        # authenticated with User accounts
-        user_db_projects = list(current_user.projects) + list(current_user.involved_in)
-        project_paths = [project.project_path for project in user_db_projects]
-        owner_ids = [project.owner_id for project in user_db_projects]
-        projects = get_projects(project_paths)
-
-    for project, owner_id in zip(projects, owner_ids):
+    # for project, owner_id in zip(projects, owner_ids):
+    for project in projects:
         try:
             project_config = project.config
-            project_config["owner_id"] = owner_id
+            # project_config["owner_id"] = owner_id  # Why is this needed?
 
             # upgrade info of v0 projects
             if project_config["version"].startswith("0"):
@@ -158,18 +148,13 @@ def api_get_projects():  # noqa: F401
 
 @bp.route("/projects/stats", methods=["GET"])
 @login_required
-def api_get_projects_stats():  # noqa: F401
+@current_user_projects
+def api_get_projects_stats(projects):  # noqa: F401
     """Get dashboard statistics of all projects"""
 
     stats_counter = {"n_in_review": 0, "n_finished": 0, "n_setup": 0}
 
-    if current_app.config.get("LOGIN_DISABLED", False):
-        project_paths = None
-    else:
-        user_db_projects = list(current_user.projects) + list(current_user.involved_in)
-        project_paths = [project.project_path for project in user_db_projects]
-
-    for project in get_projects(project_paths):
+    for project in projects:
         project_config = project.config
 
         # upgrade info of v0 projects
@@ -188,9 +173,7 @@ def api_get_projects_stats():  # noqa: F401
         except Exception:
             stats_counter["n_setup"] += 1
 
-    response = jsonify({"result": stats_counter})
-
-    return response
+    return jsonify({"result": stats_counter})
 
 
 @bp.route("/projects/info", methods=["POST"])
@@ -231,7 +214,6 @@ def api_init_project():  # noqa: F401
 @bp.route("/projects/<project_id>/upgrade_if_old", methods=["GET"])
 @login_required
 @project_authorization
-@project_from_id
 def api_upgrade_project_if_old(project):
     """Get upgrade project if it is v0.x"""
 
@@ -249,7 +231,6 @@ def api_upgrade_project_if_old(project):
 @bp.route("/projects/<project_id>/info", methods=["GET"])
 @login_required
 @project_authorization
-@project_from_id
 def api_get_project_info(project):  # noqa: F401
     """"""
     project_config = project.config
@@ -274,7 +255,6 @@ def api_get_project_info(project):  # noqa: F401
 @bp.route("/projects/<project_id>/info", methods=["PUT"])
 @login_required
 @project_authorization
-@project_from_id
 def api_update_project_info(project):  # noqa: F401
     """Update project info"""
 
@@ -331,7 +311,6 @@ def api_demo_data_project():  # noqa: F401
 @bp.route("/projects/<project_id>/data", methods=["POST", "PUT"])
 @login_required
 @project_authorization
-@project_from_id
 def api_upload_data_to_project(project):  # noqa: F401
     """"""
 
@@ -467,7 +446,6 @@ def api_upload_data_to_project(project):  # noqa: F401
 @bp.route("/projects/<project_id>/data", methods=["GET"])
 @login_required
 @project_authorization
-@project_from_id
 def api_get_project_data(project):  # noqa: F401
     """"""
 
@@ -492,7 +470,6 @@ def api_get_project_data(project):  # noqa: F401
 @bp.route("/projects/<project_id>/dataset_writer", methods=["GET"])
 @login_required
 @project_authorization
-@project_from_id
 def api_list_dataset_writers(project):
     """List the name and label of available dataset writer"""
 
@@ -539,7 +516,6 @@ def api_list_dataset_writers(project):
 @bp.route("/projects/<project_id>/search", methods=["GET"])
 @login_required
 @project_authorization
-@project_from_id
 def api_search_data(project):  # noqa: F401
     """Search for papers"""
     q = request.args.get("q", default=None, type=str)
@@ -596,7 +572,6 @@ def api_search_data(project):  # noqa: F401
 @bp.route("/projects/<project_id>/labeled", methods=["GET"])
 @login_required
 @project_authorization
-@project_from_id
 def api_get_labeled(project):  # noqa: F401
     """Get all papers classified as labeled documents"""
 
@@ -697,7 +672,6 @@ def api_get_labeled(project):  # noqa: F401
 @bp.route("/projects/<project_id>/labeled_stats", methods=["GET"])
 @login_required
 @project_authorization
-@project_from_id
 def api_get_labeled_stats(project):  # noqa: F401
     """Get all papers classified as prior documents"""
 
@@ -732,7 +706,6 @@ def api_get_labeled_stats(project):  # noqa: F401
 @bp.route("/projects/<project_id>/prior_random", methods=["GET"])
 @login_required
 @project_authorization
-@project_from_id
 def api_random_prior_papers(project):  # noqa: F401
     """Get a selection of random records.
 
@@ -881,7 +854,6 @@ def api_list_algorithms():
 @bp.route("/projects/<project_id>/algorithms", methods=["GET"])
 @login_required
 @project_authorization
-@project_from_id
 def api_get_algorithms(project):  # noqa: F401
     default_payload = {
         "model": DEFAULT_MODEL,
@@ -911,7 +883,6 @@ def api_get_algorithms(project):  # noqa: F401
 @bp.route("/projects/<project_id>/algorithms", methods=["POST"])
 @login_required
 @project_authorization
-@project_from_id
 def api_set_algorithms(project):  # noqa: F401
     # TODO@{Jonathan} validate model choice on server side
     ml_model = request.form.get("model", None)
@@ -944,7 +915,6 @@ def api_set_algorithms(project):  # noqa: F401
 @bp.route("/projects/<project_id>/start", methods=["POST"])
 @login_required
 @project_authorization
-@project_from_id
 def api_start(project):  # noqa: F401
     """Start training of first model or simulation."""
 
@@ -1028,7 +998,6 @@ def api_start(project):  # noqa: F401
 @bp.route("/projects/<project_id>/status", methods=["GET"])
 @login_required
 @project_authorization
-@project_from_id
 def api_get_status(project):  # noqa: F401
     """Check the status of the review"""
 
@@ -1054,7 +1023,6 @@ def api_get_status(project):  # noqa: F401
 @bp.route("/projects/<project_id>/status", methods=["PUT"])
 @login_required
 @project_authorization
-@project_from_id
 def api_status_update(project):
     """Update the status of the review.
 
@@ -1139,7 +1107,6 @@ def api_import_project():
 @bp.route("/projects/<project_id>/export_dataset", methods=["GET"])
 @login_required
 @project_authorization
-@project_from_id
 def api_export_dataset(project):
     """Export dataset with relevant/irrelevant labels"""
 
@@ -1227,7 +1194,6 @@ def api_export_dataset(project):
 @bp.route("/projects/<project_id>/export_project", methods=["GET"])
 @login_required
 @project_authorization
-@project_from_id
 def export_project(project):
     """Export the project file.
 
@@ -1334,7 +1300,6 @@ def _get_labels(state_obj, priors=False):
 @bp.route("/projects/<project_id>/progress", methods=["GET"])
 @login_required
 @project_authorization
-@project_from_id
 def api_get_progress_info(project):  # noqa: F401
     """Get progress statistics of a project"""
 
@@ -1349,7 +1314,6 @@ def api_get_progress_info(project):  # noqa: F401
 @bp.route("/projects/<project_id>/progress_density", methods=["GET"])
 @login_required
 @project_authorization
-@project_from_id
 def api_get_progress_density(project):
     """Get progress density of a project"""
 
@@ -1403,7 +1367,6 @@ def api_get_progress_density(project):
 @bp.route("/projects/<project_id>/progress_recall", methods=["GET"])
 @login_required
 @project_authorization
-@project_from_id
 def api_get_progress_recall(project):
     """Get cumulative number of inclusions by ASReview/at random"""
 
@@ -1445,7 +1408,6 @@ def api_get_progress_recall(project):
 @bp.route("/projects/<project_id>/record/<doc_id>", methods=["POST", "PUT"])
 @login_required
 @project_authorization
-@project_from_id
 def api_classify_instance(project, doc_id):  # noqa: F401
     """Label item
 
@@ -1501,7 +1463,6 @@ def api_classify_instance(project, doc_id):  # noqa: F401
 @bp.route("/projects/<project_id>/get_document", methods=["GET"])
 @login_required
 @project_authorization
-@project_from_id
 def api_get_document(project):  # noqa: F401
     """Retrieve documents in order of review.
 
@@ -1551,7 +1512,6 @@ def api_get_document(project):  # noqa: F401
 @bp.route("/projects/<project_id>/delete", methods=["DELETE"])
 @login_required
 @project_authorization
-@project_from_id
 def api_delete_project(project):  # noqa: F401
     """"""
